requirements.txt
langgraph>=0.0.10
langchain-google-vertexai>=0.0.5
langchain-core
langchain-community
pydantic>=2.0.0
typer>=0.9.0
python-dotenv
rich
mcp[cli]
langsmith
opentelemetry-api
opentelemetry-sdk
google-cloud-logging
networkx>=3.0
httpx
pytest
pytest-asyncio
fastapi
uvicorn

# Phase 2 Dependencies
asyncpg==0.29.0
google-cloud-tasks==2.16.3
langgraph-checkpoint-postgres==1.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
